暂时跳转回release通道，14.20版本。 15DEV版本需要自行构建，较为麻烦，且由于我们前期工作很少接触到更新的内容。

+++ Server/MainResource/Factory/Player/IPlayer.cs
@@ -7,7 +7,7 @@ namespace MainResource.Factory.Player;
 /// </summary>
 public class IPlayer : AltV.Net.Elements.Entities.Player
 {
-    public IPlayer(ICore core, IntPtr nativePointer, uint id) : base(core, nativePointer, id)
+    public IPlayer(ICore core, IntPtr nativePointer, ushort id) : base(core, nativePointer, id)
     {
     }
