+++ Server/MainResource/Factory/Player/IPlayerFactory.cs
@@ -4,7 +4,7 @@ namespace MainResource.Factory.Player;
 
 public class IPlayerFactory : IEntityFactory<IPlayer>
 {
-    public IPlayer Create(ICore core, IntPtr entityPointer, uint id)
+    public IPlayer Create(ICore core, IntPtr entityPointer, ushort id)
     {
         return new IPlayer(core, entityPointer, id);
     }
